Fixed property functions issues
